+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/EventLifetimeController.java
@@ -2,7 +2,11 @@ package nz.ac.canterbury.seng302.portfolio.controller;
 
 import nz.ac.canterbury.seng302.portfolio.model.Event;
 import nz.ac.canterbury.seng302.portfolio.service.EventService;
+import nz.ac.canterbury.seng302.portfolio.service.PermissionService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,6 +21,8 @@ public class EventLifetimeController {
     @Autowired
     private EventService eventService;
 
+    @Autowired
+    private PermissionService permissionService;
 
     /**
      * Tries to save the new event to the database
@@ -25,9 +31,12 @@ public class EventLifetimeController {
     @PostMapping("/add-event")
     public String projectSave(
             @ModelAttribute("event") Event event,
+            @AuthenticationPrincipal AuthState principal,
             Model model
     ) {
-        eventService.addEvent(event);
+        if (permissionService.isValid(principal, model)) {
+            eventService.addEvent(event);
+        }
         return "redirect:/details";
     }
 
@@ -39,9 +48,12 @@ public class EventLifetimeController {
      * @return project detail page
      */
     @GetMapping("/delete-event/{id}")
-    public String sprintRemove(@PathVariable("id") Integer id, Model model) {
-        eventService.removeEvent(id);
-
+    public String sprintRemove(@PathVariable("id") Integer id,
+                               @AuthenticationPrincipal AuthState principal,
+                               Model model) {
+        if (permissionService.isValid(principal, model)) {
+            eventService.removeEvent(id);
+        }
         /* Return the name of the Thymeleaf template */
         return "redirect:/details";
     }
